improved performance
The main issue is that a page with either a lot of comments or in
general with many dom nodes takes forever to traverse.

Therefore this is split via timeouts into smaller tasks
which prevents the UI from freezing.

After some performance tests it takes about 2-3 seconds for
16000 dom nodes.

This should be sufficient because the website is traversed from the
top to the bottum, therefore the user may not even notice this.